% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs.R
\docType{class}
\name{AFS}
\alias{AFS}
\title{AFS class to interact with data on AFS}
\format{An object of \code{\link{R6Class}} with methods to interact with AFS.}
\usage{
AFS
}
\description{
AFS class to interact with data on AFS
}
\section{Fields}{

\describe{
\item{\code{connected}}{Checks if connected to AFS.}

\item{\code{signin}}{Sign in to AFS with credentials}

\item{\code{logout}}{Logout of AFS (remove all tokens)}

\item{\code{get_tokens}}{Get the current tokens held in the cache.}

\item{\code{get_error}}{Get the last error.}

\item{\code{print_tokens}}{Print the tokens table in a few formats.}
}}
\examples{
AFS$new(cell='northstar.dartmouth.edu')
}
\keyword{datasets}

