% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{afs_load_data}
\alias{afs_load_data}
\title{Load data}
\usage{
afs_load_data(data, uppercase = TRUE)
}
\arguments{
\item{data}{name of dataset}

\item{uppercase}{Set the names all to uppercase.}
}
\value{
data.table or errors if can't reach AFS
}
\description{
Load data.  If can't connect to data, returns and empty data.table.
}

