% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R, R/set_options.R
\name{afs_set_opts}
\alias{afs_set_opts}
\title{Set user preferences}
\usage{
afs_set_opts(path = NULL, tracker = NULL)

afs_set_opts(path = NULL, tracker = NULL)
}
\arguments{
\item{path}{Path to default AFS directory}

\item{tracker}{Path to tracking file}

\item{path}{Path to default AFS directory}

\item{tracker}{Path to tracking file}
}
\description{
Set user preferences.  To make persistent, stored in inst/user

Set user preferences.  To make persistent, stored in inst/user
}

