% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_key.R
\name{create_data_key_template}
\alias{create_data_key_template}
\title{Create data_key}
\usage{
create_data_key_template(tracker = file.path(get_afs(), "file_tracker.txt"))
}
\arguments{
\item{path}{path to file tracking document}
}
\value{
data.table
}
\examples{
\dontrun{
  data_key <- create_data_key_template()
}
}

