% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{data_info}
\alias{data_info}
\title{data_info}
\usage{
data_info(path = afs, files = c(datafiles, docfiles))
}
\arguments{
\item{path}{AFS path to root directory (defaults to Lixi's folder)}

\item{files}{Files to gather info about (defaults to 'datafiles' and 'docfiles')}
}
\value{
A data.table \itemize{
  \item size: size of file in kb
  \item modified: date last modified
  \item status_change: date of last change in file status
  \item accessed: date the file was last accessed
  \item lastmod: difference in days since last modification from current date.
  \item filetype: file extension
}
}
\details{
\describe{
   Gather information about data files stored on afs.
}
}

