% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{directory_vis}
\alias{directory_vis}
\title{directory_vis}
\usage{
directory_vis(finfo, add_root = TRUE, ...)
}
\arguments{
\item{finfo}{data.table(data.frame) returned from \code{data_info}}

\item{add_root}{logical whether to add root node to join all directories (default to TRUE).}

\item{...}{Passed to plot.igraph}
}
\value{
invisibly returns the graph.
}
\description{
\description{
 Makes an igraph visualization of the file-directory structure.
}
}

