% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_info.R
\name{file_info}
\alias{file_info}
\title{file_info}
\usage{
file_info(path = get_afs(), files, fixed = TRUE, rnames = NULL)
}
\arguments{
\item{path}{AFS path to root directory (defaults to Lixi's folder)}

\item{files}{Files to gather info about (get with process_tracker)}

\item{fixed}{Match patterns exactly (surround with regexp anchors)}

\item{rnames}{R data names to match to files (if NULL then new names are made)}
}
\value{
A data.table \itemize{
  \item size: size of file in kb
  \item modified: date last modified
  \item status_change: date of last change in file status
  \item accessed: date the file was last accessed
  \item lastmod: difference in days since last modification from current date.
  \item filetype: file extension
}
}
\description{
Gather information about data files stored on afs.  Throws error if
there is no matching master file for R data.  This happens if 'rnames'
is defined and the corresponding value in 'files' can't be located.
}

