% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{nonEmpty}
\alias{nonEmpty}
\title{remove nulls/empty values from list}
\usage{
nonEmpty(lst)
}
\arguments{
\item{lst}{list}
}
\value{
list w/o nulls/empty values
}
\description{
Remove empty/null values from list
}

