% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_key.R
\name{set_key_names}
\alias{set_key_names}
\title{Update R file names associated with master files.}
\usage{
set_key_names(rnames, filenames, key = getOption("afs.key"))
}
\arguments{
\item{rnames}{New R data names}

\item{filenames}{Corresponding master file names.}

\item{key}{Data key to update (getOption('afs.key))}
}
\description{
Update R file names associated with master files.
}

