% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_in.R
\name{signin}
\alias{signin}
\title{Sign in to AFS, if options aren't supplied, try interactive
... ignored for now, but can be other options to klog}
\usage{
signin(user, pwd, cell, ...)
}
\description{
Sign in to AFS, if options aren't supplied, try interactive
... ignored for now, but can be other options to klog
}

