% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_key.R
\name{update_key}
\alias{update_key}
\title{Update data_key by processing the tracking document and renaming/adding
files to data_key.}
\usage{
update_key(path = getOption("afs.path"), tracker = getOption("afs.tracker"),
  data = get_key(), save_key = TRUE, outfile = NULL)
}
\arguments{
\item{path}{Path to AFS root directory (getOption('afs.path))}

\item{tracker}{tracking file (getOption('afs.tracker'))}

\item{data}{Data key (getOption('afs.key')}

\item{save_key}{Save the returned key (default to "pkg/extdata/data_key.rda")}

\item{outfile}{Where to save key, only used if save_key is TRUE.}
}
\description{
Update data_key by processing the tracking document and renaming/adding
files to data_key.
}

