% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sign_in.R
\name{validate_input}
\alias{validate_input}
\title{Validate user input}
\usage{
validate_input(prompt, expect = "character", allow_blank = FALSE)
}
\arguments{
\item{prompt}{Message to display to user}

\item{expect}{Type expected as response}

\item{allow_blank}{Allow empty responses}
}
\description{
Validate user input to be correct type
}

