% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{afs_set_opts}
\alias{afs_set_opts}
\title{Set user preferences}
\usage{
afs_set_opts(cell, tracker, userdir, key)
}
\arguments{
\item{cell}{AFS cell}

\item{tracker}{Path to tracking file}

\item{userdir}{Path to users directory on AFS (after the cell)}

\item{key}{Data key (default to 'data_key.rda' in 'extdata')}
}
\description{
Set user preferences.  To make persistent, stored in inst/user
}

