% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files.R
\name{copy_files}
\alias{copy_files}
\title{Copy files from AFS}
\usage{
copy_files(ext, afs, outdir = "temp", dkey = sync.afs::data_key, ...)
}
\arguments{
\item{ext}{File extensions (character vector)}

\item{afs}{An AFS object with path set to user root directory}

\item{outdir}{Name of directory to copy files to}

\item{dkey}{Data key}

\item{...}{Passed to \code{file.copy}}
}
\value{
TRUE if all files copied successfully.
}
\description{
Copy files from AFS given file extensions.
}

