% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_key.R
\name{create_data_key_template}
\alias{create_data_key_template}
\title{Create data_key (as data.table) to store files/locations/info for quicker access}
\usage{
create_data_key_template(path = getOption("afs.path"),
  tracker = getOption("afs.tracker"))
}
\arguments{
\item{path}{Path to root directory.}

\item{tracker}{Path to data tracking file (default getOption('afs.tracker'))}
}
\value{
data.table
}
\description{
Create data_key (as data.table) to store files/locations/info for quicker access
}
\examples{
\dontrun{
  data_key <- create_data_key_template()
}
}

