% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_key.R
\name{create_data_key_template}
\alias{create_data_key_template}
\title{Create data_key}
\usage{
create_data_key_template(path = sync.afs::get_afs())
}
\arguments{
\item{path}{path to afs directory}
}
\value{
data.table
}
\examples{
\dontrun{
  data_key <- create_data_key_template()
}
}
\keyword{internal}

