% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{data_vis}
\alias{data_vis}
\title{Wrapper to visualize the location of files in the AFS directory.
Processes a tracking document to get files prior to passing to directory_vis.}
\usage{
data_vis(afs, tracker = getOption("afs.tracker"), add_root = TRUE, ...)
}
\arguments{
\item{afs}{An AFS object with path set to user root}

\item{tracker}{Path to tracker file (default getOption('afs.tracker'))}

\item{add_root}{Logical to add root to graph}

\item{...}{passed to plot.igraph}

\item{path}{Path to root AFS folder (has default)}
}
\value{
Returns the igraph invisibly.
}
\description{
Wrapper to visualize the location of files in the AFS directory.
Processes a tracking document to get files prior to passing to directory_vis.
}
\examples{
\dontrun{

   ## Defaults to Lixi's root directory and tracked files.
   data_vis()

}
}

