% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{data_vis}
\alias{data_vis}
\title{data_vis}
\usage{
data_vis(tracker = file.path(get_afs(), "file_tracker.txt"),
  path = get_afs(), add_root = TRUE, ...)
}
\arguments{
\item{tracker}{Path to tracker file on AFS (has default)}

\item{path}{Path to root AFS folder (has default)}

\item{add_root}{Logical to add root to graph}

\item{...}{passed to plot.igraph}
}

