% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/vis.R
\name{data_vis}
\alias{data_vis}
\title{data_vis}
\usage{
data_vis(tracker = file.path(get_afs(), "file_tracker.txt"),
  path = get_afs(), add_root = TRUE, ...)
}
\arguments{
\item{tracker}{Path to tracker file on AFS (has default)}

\item{path}{Path to root AFS folder (has default)}

\item{add_root}{Logical to add root to graph}

\item{...}{passed to plot.igraph}
}
\value{
Returns the igraph invisibly.
}
\description{
Wrapper to visualize the location of files in the AFS directory.
Processes a tracking document to get files prior to passing to directory_vis.
}
\examples{
\dontrun{

   ## Defaults to Lixi's root directory and tracked files.
   data_vis()

}
}

