% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_info.R
\name{file_info}
\alias{file_info}
\title{file_info}
\usage{
file_info(path = if (Sys.info()[["sysname"]] == "Linux") afs.linux else afs,
  files)
}
\arguments{
\item{path}{AFS path to root directory (defaults to Lixi's folder)}

\item{files}{Files to gather info about (get with process_tracker)}
}
\value{
A data.table \itemize{
  \item size: size of file in kb
  \item modified: date last modified
  \item status_change: date of last change in file status
  \item accessed: date the file was last accessed
  \item lastmod: difference in days since last modification from current date.
  \item filetype: file extension
}
}
\details{
Gather information about data files stored on afs.
}

