% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read master data file corresponding to a R data filename}
\usage{
get_data(dname, ...)
}
\arguments{
\item{dname}{data name in R}

\item{...}{arguments to be passed to reading function}
}
\value{
corresponding master file (prior to cleaning)
}
\examples{
\dontrun{
  pp_raw <- get_data('pp_raw')
}
}
\author{
Noah Peart
}

