% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read master data file corresponding to a R data filename}
\usage{
get_data(dname, path = get_afs(), dkey = sync.afs::data_key, ...)
}
\arguments{
\item{dname}{data name in R}

\item{path}{Path to root AFS directory}

\item{dkey}{data key to match R data files to master files (default data_key)}

\item{...}{arguments to be passed to data reading function}

\item{data_key}{Key matching R data to master files.}
}
\value{
data.table
}
\description{
Read master data associated with R data file.
}
\examples{
\dontrun{
  pp_raw <- get_data('pp_raw')
}
}

