% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracker.R
\name{process_tracker}
\alias{process_tracker}
\title{process_tracker}
\usage{
process_tracker(tracker = file.path(get_afs(), "file_tracker.txt"))
}
\arguments{
\item{tracker}{Path to file tracking file}
}
\value{
list of two elements: filenames to track, and renamed files
}
\details{
Process the file_tracker.txt document.  If files were renamd, replace with new names.
Also, if filenames are duplicated, they are removed.
}

