% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_key.R
\name{update_key}
\alias{update_key}
\title{Update the data_key with new entries/name changes}
\usage{
update_key(path = get_afs(), tracker = "file_tracker.txt",
  data_key = data_key, save_key = FALSE)
}
\arguments{
\item{path}{Path to base AFS directory from which to search (default to get_afs()).}

\item{tracker}{Name of the text file containing filenames to track (default to 'file_tracker.txt').}

\item{data_key}{The current data_key to update (default to data/data_key.rda).}

\item{save_key}{Save the key in the data directory?}
}
\keyword{internal}

