% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/update_key.R
\name{update_key}
\alias{update_key}
\title{Update the data_key with new entries/name changes}
\usage{
update_key(path = get_afs(), tracker = "file_tracker.txt",
  data = data_key, save_key = FALSE, outfile = NULL)
}
\arguments{
\item{path}{Path to base AFS directory from which to search (default to get_afs()).}

\item{tracker}{Name of the text file containing filenames to track (default to 'file_tracker.txt').}

\item{data}{The current data key to update (default to data/data_key.rda).}

\item{save_key}{Save the key in the data directory?  Use this if going to push update.}

\item{outfile}{Where to save key, only used if save_key is TRUE.}
}
\description{
Update data_key by processing the tracking document and renaming/adding
files to data_key.
}

